Fixed issue with DeviceRotation user comment field on MetaInformation. Removed MotionManager (It is not longer needed) Moved extensions to Extensions file
Issue #41
