+++ GiniVision/Classes/Extensions.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 import AVFoundation
+import ImageIO
 
 /**
  Returns an optional `UIImage` instance with the given `name` preferably from the client's bundle.
@@ -151,3 +152,125 @@ internal extension UIDevice {
     }
     
 }
+
+internal extension Collection where Iterator.Element == CFString {
+    
+    var strings: [ String ] {
+        return self.map { $0 as String }
+    }
+    
+}
+
+internal extension NSMutableDictionary {
+    
+    fileprivate var cfExifKeys: [CFString] {
+        return [
+            kCGImagePropertyExifLensMake,
+            kCGImagePropertyExifLensModel,
+            kCGImagePropertyExifISOSpeed,
+            kCGImagePropertyExifISOSpeedRatings,
+            kCGImagePropertyExifExposureTime,
+            kCGImagePropertyExifApertureValue,
+            kCGImagePropertyExifFlash,
+            kCGImagePropertyExifCompressedBitsPerPixel,
+            kCGImagePropertyExifUserComment
+        ]
+    }
+    
+    fileprivate var cfTiffKeys: [CFString] {
+        return [
+            kCGImagePropertyTIFFMake,
+            kCGImagePropertyTIFFModel,
+            kCGImagePropertyTIFFSoftware,
+            kCGImagePropertyTIFFOrientation
+        ]
+    }
+    
+    fileprivate var cfTopLevelKeys: [CFString] {
+        return [
+            kCGImagePropertyExifDictionary,
+            kCGImagePropertyTIFFDictionary
+        ]
+    }
+    
+    fileprivate var stringKeys: [String] {
+        return allKeys
+            .map { $0 as? String }
+            .flatMap { $0 }
+    }
+    
+    func set(metaInformation value: AnyObject?, forKey key: String, inSubdictionary isSubdictionary: Bool = false) {
+        // Helper method to set a value in a meta dictionary like TIFF or Exif
+        func set(_ value: AnyObject?, forKey key: String, inMetaDictionaryWithKey metaDictionaryKey: String) {
+            let metaDictionary = (self[metaDictionaryKey] as? NSMutableDictionary) ?? NSMutableDictionary()
+            metaDictionary[key] = value
+            self[metaDictionaryKey] = metaDictionary
+        }
+        
+        // Try to set the key in the current context
+        let keys = stringKeys
+        if keys.contains(key) {
+            setValue(value, forKey: key)
+        }
+        
+        // Try to set in known context
+        if !isSubdictionary {
+            if cfExifKeys.strings.contains(key) {
+                set(value, forKey: key, inMetaDictionaryWithKey: kCGImagePropertyExifDictionary as String)
+            }
+            if cfTiffKeys.strings.contains(key) {
+                set(value, forKey: key, inMetaDictionaryWithKey: kCGImagePropertyTIFFDictionary as String)
+            }
+        }
+        
+        // Try to set in subdictioanies
+        let dictionaries = allValues
+            .map { $0 as? NSMutableDictionary }
+            .flatMap { $0 }
+        
+        for dictionary in dictionaries {
+            dictionary.set(metaInformation: value, forKey: key, inSubdictionary: true)
+        }
+    }
+    
+    func getMetaInformation(forKey key: String) -> AnyObject? {
+        let keys = stringKeys
+        guard keys.count > 0 else { return nil }
+        if keys.contains(key) {
+            return self[key] as AnyObject?
+        }
+        let dictionaries = allValues
+            .map { $0 as? NSMutableDictionary }
+            .flatMap { $0 }
+        
+        for dictionary in dictionaries {
+            if let value = dictionary.getMetaInformation(forKey: key) {
+                return value
+            }
+        }
+        
+        return nil
+    }
+    
+    func filterDefaultMetaInformation() {
+        let keys = [ cfExifKeys, cfTiffKeys, cfTopLevelKeys ].flatMap { $0 }.strings
+        filterMetaInformation(keys)
+    }
+    
+    /**
+     Set all values to `CFNull` if the corresponding key is not in the filtered keys array.
+     
+     - parameter filterKeys: The keys to filter for.
+     */
+    func filterMetaInformation(_ filterKeys: [String]) {
+        let keys = stringKeys
+        for key in keys {
+            if !filterKeys.contains(key) {
+                self[key] = kCFNull
+            } else if let dictionary = self[key] as? NSMutableDictionary {
+                dictionary.filterDefaultMetaInformation()
+            }
+        }
+    }
+}
+
